Solved issue of items are equal when adding
When you do tempCartItems = List.from(selectedOrder.items);, it creates a new list, but the items in the list are not deep copied. So if the items themselves are updated (like increasing the quantity), it will reflect in both selectedOrder.items and tempCartItems.
This is the reason why the compare function return true because somewhere the tempCartItems are being updated with the same quantity when selectedOrder.items also updated. They are pointing to the same memory list. A deep copy is need.